A function to create a section of the parameter input for each MAS type

A function to create a section of the parameter input for each MAS type

create_par_section(
  section_type,
  section_type_object,
  id = 1,
  par_names,
  par_lo = NULL,
  par_hi = NULL,
  par_units = NULL,
  par_phase = NULL,
  par_value,
  rec_devs = NA
)

Arguments

section_type
The type of section; string of type for example "recruitment", "growth".

section_type_object
The object created within the RMAS interface denoting the type of function used

id
The id of this parameter section; if there's only one parameter section of this type in the model, leave blank to assign the default ID of 1.

par_names
vector of the parameter names, must match the names denoted in the RMAS C++ file

par_lo
vector of the parameter lower bounds, default NULL will assign default lower bound

par_hi
vector of the parameter upper bounds, default NULL will assign default upper bound

par_units
string denoting the units of the parameter, default NULL will assume the parameter is unitless

par_phase
vector of estimation phases for this parameter, default NULL will assume this uses the MAS default

par_value
vector of values for this parameter, note this will be translated to C++ and cast into the appropriate type (for example, double, integer)

rec_devs
supply only if this parameter section is of type "recruitment"

Value

the section_type_object with the populated parameter values